chore(dragdrop): Move all macOS DataPackage integration to separate file
