+++ src/Uno.UWP/ApplicationModel/DataTransfer/Clipboard.macOS.cs
@@ -16,98 +16,12 @@ namespace Windows.ApplicationModel.DataTransfer
 
 		public static void SetContent(DataPackage content)
 		{
-			if (content is null)
-			{
-				throw new ArgumentNullException(nameof(content));
-			}
-
-			// See notes in Clipboard.Android.cs on async code usage here
-			CoreDispatcher.Main.RunAsync(
-				CoreDispatcherPriority.High,
-				async () =>
-				{
-					var data = content?.GetView();
-
-					var declaredTypes = new List<string>();
-					var pasteboard = NSPasteboard.GeneralPasteboard;
-
-					// Note that order is somewhat important here.
-					//
-					// According to the docs:
-					//    "types should be ordered according to the preference of the source application,
-					//     with the most preferred type coming first"
-					// https://developer.apple.com/documentation/appkit/nspasteboard/1533561-declaretypes?language=objc
-					//
-					// This means we want to process certain types like HTML/RTF before general plain text
-					// as they are more specific.
-					// Types are also declared before setting
-
-					// Declare types
-					if (data?.Contains(StandardDataFormats.Html) ?? false)
-					{
-						declaredTypes.Add(NSPasteboard.NSPasteboardTypeHTML);
-					}
-
-					if (data?.Contains(StandardDataFormats.Rtf) ?? false)
-					{
-						declaredTypes.Add(NSPasteboard.NSPasteboardTypeRTF);
-					}
-
-					if (data?.Contains(StandardDataFormats.Text) ?? false)
-					{
-						declaredTypes.Add(NSPasteboard.NSPasteboardTypeString);
-					}
-
-					pasteboard.DeclareTypes(declaredTypes.ToArray(), null);
-
-					// Set content
-					if (data?.Contains(StandardDataFormats.Html) ?? false)
-					{
-						var html = await data.GetHtmlFormatAsync();
-						pasteboard.SetStringForType(html ?? string.Empty, NSPasteboard.NSPasteboardTypeHTML);
-					}
-
-					if (data?.Contains(StandardDataFormats.Rtf) ?? false)
-					{
-						var rtf = await data.GetRtfAsync();
-						pasteboard.SetStringForType(rtf ?? string.Empty, NSPasteboard.NSPasteboardTypeRTF);
-					}
-
-					if (data?.Contains(StandardDataFormats.Text) ?? false)
-					{
-						var text = await data.GetTextAsync();
-						pasteboard.SetStringForType(text ?? string.Empty, NSPasteboard.NSPasteboardTypeString);
-					}
-				});
-
-			return;
+			DataPackage.SetToNativeClipboard(content);
 		}
 
 		public static DataPackageView GetContent()
 		{
-			var dataPackage = new DataPackage();
-
-			var pasteboard = NSPasteboard.GeneralPasteboard;
-
-			var clipHtml = pasteboard.GetStringForType(NSPasteboard.NSPasteboardTypeHTML);
-			if (clipHtml != null)
-			{
-				dataPackage.SetHtmlFormat(clipHtml);
-			}
-
-			var clipRtf = pasteboard.GetStringForType(NSPasteboard.NSPasteboardTypeRTF);
-			if (clipRtf != null)
-			{
-				dataPackage.SetRtf(clipRtf);
-			}
-
-			var clipText = pasteboard.GetStringForType(NSPasteboard.NSPasteboardTypeString);
-			if (clipText != null)
-			{
-				dataPackage.SetText(clipText);
-			}
-
-			return dataPackage.GetView();
+			return DataPackage.GetFromNativeClipboard();
 		}
 
 		public static void Clear()
